Adding changes for srgb

+++ Engine/Graphics/cMaterial.cpp
@@ -9,7 +9,7 @@ namespace
 	std::string m_EffectLocation;
 	uint8_t m_ConstantType;
 	std::string m_ConstantName;
-	std::vector<uint8_t> m_ConstantData;
+	std::vector<float> m_ConstantData;
 	std::string m_TextureLocation;
 }
 
@@ -31,8 +31,8 @@ eae6320::cResult LoadDataFile(const char* const i_FileName)
 	m_ConstantData.clear();
 	for (size_t i = 0; i < m_ConstantType + 1; i++)
 	{
-		m_ConstantData.push_back(*reinterpret_cast<uint8_t*>(offset));
-		offset += sizeof(uint8_t);
+		m_ConstantData.push_back(*reinterpret_cast<float*>(offset));
+		offset += sizeof(float);
 	}
 	m_TextureLocation = reinterpret_cast<char*>(offset);
 	result = eae6320::Graphics::cEffect::s_Manager.Load(m_EffectLocation, m_TempHandle);
@@ -53,10 +53,7 @@ eae6320::cResult eae6320::Graphics::cMaterial::Load(const std::string& i_FilePat
 	}
 	outputMaterial->m_EffectHandle = m_TempHandle;
 	outputMaterial->m_TextureHandle = m_TempTextureHandle;
-	outputMaterial->m_Color = { static_cast<float>(m_ConstantData[0]),
-		static_cast<float>(m_ConstantData[1]),
-		static_cast<float>(m_ConstantData[2]),
-		static_cast<float>(m_ConstantData[3]) };
+	outputMaterial->m_Color = { (m_ConstantData[0]),(m_ConstantData[1]),(m_ConstantData[2]),(m_ConstantData[3]) };
 
 	o_Material = outputMaterial;
 	return result;
